Feat: Modificacion clases View
