+++ SistemaTaller/src/view/ServicioView.java
@@ -1,5 +1,7 @@
 package view;
 
 public class ServicioView {
-    
+    public void gestionServicios() {
+    }
+
 }
